requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dj-database-url==2.0.0
Django==4.2.1
django-environ==0.10.0
django-redis==5.2.0
django-render-block==0.9.2
django-templated-email==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.5
requests==2.30.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==2.0.2
zope.interface==6.0
